refactor(analysis-types): extract determinism baseline DTO
